+++ core/modelspec/uses.js
@@ -6,7 +6,7 @@
  */
 import {throwIf} from '@xh/hoist/utils/js';
 import {ModelPublishMode, ModelSpec} from './ModelSpec';
-import {isFunction} from 'lodash';
+import {isFunction, isString} from 'lodash';
 
 /**
  * Returns a ModelSpec to define how a functional HoistComponent should source its primary backing
@@ -34,6 +34,7 @@ import {isFunction} from 'lodash';
  *      construct an instance on-demand. Selector must be a HoistModel Class.
  * @param {(boolean|function)} [flags.createDefault] - true to create a model if none provided.
  *      Selector must be a HoistModel Class, or a custom function may be provided for this argument.
+ * @param {boolean} [flags.optional] - true to specify a model that is optional.  Default false.
  * @returns {ModelSpec}
  */
 export function uses(
@@ -41,12 +42,18 @@ export function uses(
         fromContext = true,
         publishMode = ModelPublishMode.DEFAULT,
         createFromConfig = true,
-        createDefault = false
+        createDefault = false,
+        optional = false
     } = {}) {
-    return new UsesSpec(selector, fromContext, publishMode, createFromConfig, createDefault);
+    return new UsesSpec(selector, fromContext, publishMode, createFromConfig, createDefault, optional);
 }
 
 
+/**
+ * Ensure an object is a ModelSelector, or throw.
+ *
+ * @param {Object} s
+ */
 export function ensureIsSelector(s) {
     throwIf(
         !isFunction(s) && s !== '*',
@@ -55,6 +62,26 @@ export function ensureIsSelector(s) {
     );
 }
 
+/**
+ * Format a ModelSelector for debugging purposes.
+ *
+ * @param {ModelSelector} selector
+ * @returns {string}
+ */
+export function formatSelector(selector) {
+    if (isString(selector)) return selector;
+    if (isFunction(selector)) {
+        if (selector.isHoistModel) return selector.name;
+        try {
+            selector = selector();
+        } catch (e) {
+        }
+        if (selector.isHoistModel) return '() => ' + selector.name;
+    }
+    return '[Selector]';
+}
+
+
 /** @private */
 export class UsesSpec extends ModelSpec {
 
@@ -62,8 +89,8 @@ export class UsesSpec extends ModelSpec {
     createFromConfig;
     createDefault;
 
-    constructor(selector, fromContext, publishMode, createFromConfig, createDefault) {
-        super(fromContext, publishMode);
+    constructor(selector, fromContext, publishMode, createFromConfig, createDefault, optional) {
+        super(fromContext, publishMode, optional);
 
         ensureIsSelector(selector);