+++ CHANGELOG.md
@@ -18,6 +18,10 @@
   the new features described above are all "opt-in" and should be fully backward compatible with
   existing application code.
 
+* Improvements to `uses()` directive to `HoistComponent`.  Hoist will now more cleanly alert
+  if a model specified with this directive cannot be resolved.   A new `optional` config
+  (default false) supports the case of components that support optional models.
+
 ### 💥 Breaking Changes
    Dashboard widgets no longer receive the `viewModel` prop.  Access to the `DashViewModel` within a
    widget should be obtained using either the lookup decorator (i.e. `@lookup(DashViewModel)`)